requirements.txt
PyPDF2>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
google-search-results>=2.4.2
openai>=1.0.0
PyYAML>=6.0.1
pytest>=7.4.0
pytest-mock
responses
